% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{get_missingness_info}
\alias{get_missingness_info}
\title{Get native missingness summaries}
\usage{
get_missingness_info(fit)
}
\arguments{
\item{fit}{A \code{rajive_incomplete} object.}
}
\value{
A list of block, sample, feature, cell, and count summaries.
}
\description{
Get native missingness summaries
}
