% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{plot_variance_explained}
\alias{plot_variance_explained}
\title{Bar chart of variance explained}
\usage{
plot_variance_explained(ajive_output, blocks)
}
\arguments{
\item{ajive_output}{List returned by \code{\link{Rajive}}.}

\item{blocks}{List of data matrices (the same list passed to
\code{\link{Rajive}}).}
}
\value{
A \code{ggplot2} object.
}
\description{
Produces a stacked bar chart showing the proportion of total variance
explained by the joint, individual, and residual components for each
data block.
}
\examples{
\donttest{
n <- 30; pks <- c(40, 30)
Y <- ajive.data.sim(K = 2, rankJ = 2, rankA = c(5, 4), n = n,
                    pks = pks, dist.type = 1)
data.ajive <- Y$sim_data
res <- Rajive(data.ajive, c(5, 4))
plot_variance_explained(res, data.ajive)
}

}
