% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw.R
\name{print.jackstraw_rajive}
\alias{print.jackstraw_rajive}
\title{Print method for jackstraw_rajive objects}
\usage{
\method{print}{jackstraw_rajive}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"jackstraw_rajive"}.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Displays a concise summary table of significance results for each block and
joint component.
}
