% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{print.rajive}
\alias{print.rajive}
\title{Print method for rajive objects}
\usage{
\method{print}{rajive}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rajive"} returned by \code{\link{Rajive}}.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Displays a concise summary of the RaJIVE decomposition: number of blocks,
estimated joint rank, and individual rank for each data block.
}
\examples{
\donttest{
n <- 30; pks <- c(40, 30)
Y <- ajive.data.sim(K = 2, rankJ = 2, rankA = c(5, 4), n = n,
                    pks = pks, dist.type = 1)
res <- Rajive(Y$sim_data, c(5, 4))
print(res)
}

}
