% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rajiveplus-package.R
\docType{package}
\name{rajiveplus-package}
\alias{rajiveplus}
\alias{rajiveplus-package}
\title{rajiveplus: Robust Angle Based Joint and Individual Variation Explained}
\description{
Blazing fast re-implementation of the RAJIVE algorithm for joint and individual variation explained in multi-view data. This package provides a user-friendly interface to perform RAJIVE analysis, along with tools for visualization and interpretation of results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mdmanurung.github.io/rajiveplus/}
  \item Report bugs at \url{https://github.com/mdmanurung/rajiveplus/issues}
}

}
\author{
\strong{Maintainer}: Mikhael Manurung \email{m.d.manuurng@lumc.nl}

Authors:
\itemize{
  \item Mikhael Manurung \email{m.d.manuurng@lumc.nl}
}

}
