% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rajiveplus-package.R
\docType{package}
\name{rajiveplus-package}
\alias{rajiveplus}
\alias{rajiveplus-package}
\title{rajiveplus: Robust Angle Based Joint and Individual Variation Explained}
\description{
A robust alternative to the aJIVE (angle based Joint and Individual Variation Explained) method (Feng et al 2018: \doi{10.1016/j.jmva.2018.03.008}) for the estimation of joint and individual components in the presence of outliers in multi-source data. It decomposes the multi-source data into joint, individual and residual (noise) contributions. The decomposition is robust to outliers and noise in the data. The method is illustrated in Ponzi et al (2021) \href{https://arxiv.org/abs/2101.09110}{arXiv:2101.09110}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mdmanurung.github.io/rajiveplus/}
  \item Report bugs at \url{https://github.com/mdmanurung/rajiveplus/issues}
}

}
\author{
\strong{Maintainer}: Erica Ponzi \email{erica.ponzi@medisin.uio.no}

Authors:
\itemize{
  \item Abhik Ghosh \email{abhik.ghosh.stat@gmail.com}
}

}
